+++ src/rendering/index.ts
@@ -63,3 +63,4 @@ export * from './View';
 export * from './filters/Filter';
 export * from './filters/BlurFilter';
 export * from './filters/ColorFilter';
+export * from './filters/ShaderFilter';